remove outline, add drag drop functionality to content blocks
