+++ assets/src/components/resource/DragHandle.tsx
@@ -2,14 +2,15 @@ import * as React from 'react';
 
 export interface DragHandleProps {
   hidden?: boolean;
+  style?: any;
 }
 
 export const DragHandle = (props: DragHandleProps) => {
 
-  const { hidden = false } = props;
+  const { hidden = false, style = {} } = props;
 
   return (
-    <div className={`drag-handle ${hidden ? 'invisible' : ''}`}>
+    <div className={`drag-handle ${hidden ? 'invisible' : ''}`} style={style}>
       <div className="grip" />
     </div>
   );
