+++ assets/src/components/resource/Editors.tsx
@@ -1,18 +1,30 @@
 import * as Immutable from 'immutable';
-import React from 'react';
-import { ResourceContent, Activity, ResourceType, ActivityPurposes, ContentPurposes } from 'data/content/resource';
+import React, { useState } from 'react';
+import {
+  ResourceContent, Activity, ResourceType, ActivityPurposes, ContentPurposes,
+  ActivityReference, StructuredContent,
+} from 'data/content/resource';
 import { ActivityEditorMap, EditorDesc } from 'data/content/editors';
 import { UnsupportedActivity } from './UnsupportedActivity';
 import { getToolbarForResourceType } from './toolbar';
 import { StructuredContentEditor } from 'components/content/StructuredContentEditor';
-import { ResourceContentFrame } from 'components/content/ResourceContentFrame';
 import { ProjectSlug, ResourceSlug } from 'data/types';
 import { TestModeHandler, defaultState } from './TestModeHandler';
+import { Purpose } from '../content/Purpose';
+import { DeleteButton } from '../misc/DeleteButton';
+import { EditLink } from '../misc/EditLink';
+import * as Persistence from 'data/persistence/activity';
+
+import "./Editors.scss"
+import { getContentDescription, toSimpleText } from 'data/content/utils';
+import { DragHandle } from './DragHandle';
+import { classNames } from 'utils/classNames';
 
 export type EditorsProps = {
   editMode: boolean,              // Whether or not we can edit
   content: Immutable.List<ResourceContent>,     // Content of the resource
   onEdit: (content: ResourceContent, index: number) => void,
+  onEditContentList: (content: Immutable.List<ResourceContent>) => void,
   onRemove: (index: number) => void,
   editorMap: ActivityEditorMap,   // Map of activity types to activity elements
   graded: boolean,
@@ -21,12 +33,61 @@ export type EditorsProps = {
   resourceSlug: ResourceSlug,
 };
 
+interface ActivityPayload {
+  type: 'ActivityPayload';
+  id: string;
+  activity: Activity;
+  reference: ActivityReference;
+  project: ProjectSlug;
+}
+
+interface UnknownPayload {
+  type: 'UnknownPayload';
+  id: string;
+  data: any;
+}
+
+type DragPayload = StructuredContent | ActivityPayload | UnknownPayload;
+
+// @ts-ignore
+const DropTarget = ({ id, index, onDrop }) => {
+  const [hovered, setHovered] = useState(false);
+
+  const handleDragEnter = (e: React.DragEvent<HTMLDivElement>) => setHovered(true);
+  const handleDragLeave = (e: React.DragEvent<HTMLDivElement>) => setHovered(false);
+  const handleDrop = (e: React.DragEvent<HTMLDivElement>) => {
+    e.preventDefault();
+    setHovered(false);
+    onDrop(e, index);
+  };
+  const handleOver = (e: React.DragEvent<HTMLDivElement>) => {
+    e.stopPropagation();
+    e.preventDefault();
+  };
+
+  return (
+    <div key={id + '-drop'}
+      className={classNames(['drop-target ', hovered ? 'hovered' : ''])}
+      onDragEnter={handleDragEnter}
+      onDragLeave={handleDragLeave}
+      onDrop={handleDrop}
+      onDragOver={handleOver}>
+    </div>
+  );
+};
+
+const getFriendlyName = (item: ActivityReference, editorMap: ActivityEditorMap,
+  activities: Immutable.Map<string, Activity>) => {
+
+  const activity = activities.get(item.activitySlug);
+  return editorMap[(activity as any).typeSlug].friendlyName;
+};
 
 // The list of editors
 export const Editors = (props: EditorsProps) => {
 
   const { editorMap, editMode, graded,
-    content, activities, projectSlug, resourceSlug } = props;
+    content, activities, projectSlug, resourceSlug, onEditContentList } = props;
 
   // Factory for creating top level editors, for things like structured
   // content or referenced activities
@@ -92,44 +153,228 @@ export const Editors = (props: EditorsProps) => {
 
   };
 
+  const [assisstive, setAssisstive] = useState('');
+
+  const onFocus = (index: number) => {
+    const item = content.get(index) as ResourceContent;
+    const desc = item.type === 'content'
+      ? getContentDescription(item)
+      : getFriendlyName(item, editorMap, activities);
+
+    setAssisstive(
+      `Listbox. ${index + 1} of ${content.size}. ${desc}.`);
+  };
+
+  const onMove = (index: number, up: boolean) => {
+
+    if (index === 0 && up) return;
+
+    const item = content.get(index) as ResourceContent;
+    const inserted = content
+      .remove(index)
+      .insert(index + (up ? -1 : 1), item as any);
+
+    onEditContentList(inserted);
+
+    const newIndex = inserted.findIndex(c => c.id === item.id);
+    const desc = item.type === 'content'
+      ? 'Content' : getFriendlyName(item, editorMap, activities);
+
+    setAssisstive(
+      `Listbox. ${newIndex + 1} of ${content.size}. ${desc}.`);
+  };
+
+  const onDrop = (e: React.DragEvent<HTMLDivElement>, index: number) => {
+    if (editMode) {
+      const data = e.dataTransfer.getData('application/x-oli-resource-content');
+
+      if (data) {
+        const droppedContent = JSON.parse(data) as DragPayload;
+
+        const sourceIndex = content.findIndex(c => c.id === droppedContent.id);
+
+        if (sourceIndex === -1) {
+
+          // This is a cross window drop, we insert it but have to have to
+          // ensure that for activities that we create a new activity for
+          // tied to this project
+          if (droppedContent.type === 'ActivityPayload') {
+
+            if (droppedContent.project !== projectSlug) {
+
+              Persistence.create(
+                droppedContent.project,
+                droppedContent.activity.typeSlug,
+                droppedContent.activity.model)
+              .then((result: Persistence.Created) => {
+                onEditContentList(content.insert(index, droppedContent.reference));
+              });
+
+            } else {
+              onEditContentList(content.insert(index, droppedContent.reference));
+            }
+
+
+          } else if (droppedContent.type === 'content') {
+            onEditContentList(content.insert(index, droppedContent));
+          } else {
+            onEditContentList(content.insert(index, droppedContent.data));
+          }
+
+          return;
+
+        }
+        if (sourceIndex > -1) {
+          // Handle a same window drag and drop
+          const adjusted = sourceIndex < index ? index - 1 : index;
+
+          let toInsert;
+          if (droppedContent.type === 'ActivityPayload') {
+            toInsert = droppedContent.reference;
+          } else if (droppedContent.type === 'content') {
+            toInsert = droppedContent;
+          } else {
+            toInsert = droppedContent.data;
+          }
+
+          const reordered = content.remove(sourceIndex).insert(adjusted, toInsert);
+          onEditContentList(reordered);
+          return;
+        }
+      }
+
+      // Handle a drag and drop from VSCode
+      const text = e.dataTransfer.getData('codeeditors');
+      if (text) {
+        try {
+          const json = JSON.parse(text);
+          const parsedContent = JSON.parse(json[0].content);
+
+          // Remove it if we find the same identified content item
+          const inserted = content
+            .filter(c => parsedContent.id !== c.id)
+            // Then insert it
+            .insert(index, parsedContent);
+
+            onEditContentList(inserted);
+        } catch (err) {
+
+        }
+      }
+
+    }
+  };
+
 
   const editors = content.map((c, index) => {
 
     const onEdit = (u : ResourceContent) => props.onEdit(u, index);
     const onRemove = () => props.onRemove(index);
+    const onEditPurpose = (purpose: string) => {
+      const u = Object.assign(c, { purpose });
+      props.onEdit(u, index);
+    };
 
     const [editor, label] = createEditor(c, onEdit);
 
     const editingLink = c.type === 'activity-reference'
       ? `/project/${projectSlug}/resource/${resourceSlug}/activity/${c.activitySlug}` : undefined;
 
+    const link = editingLink !== undefined
+      ? (
+          <EditLink href={editingLink}/>
+        )
+      : null;
+
     const purposes = c.type === 'activity-reference'
       ? ActivityPurposes : ContentPurposes;
 
+    let dragPayload : DragPayload;
+    if (c.type === 'content') {
+      dragPayload = c;
+    } else if (activities.has(c.activitySlug)) {
+      const activity = activities.get(c.activitySlug);
+      dragPayload = {
+        type: 'ActivityPayload',
+        id: c.id,
+        reference: c,
+        activity: activity as Activity,
+        project: projectSlug,
+      } as ActivityPayload;
+    } else {
+      dragPayload = { type: 'UnknownPayload', data: c, id: c.id };
+    }
+
+    const handleDragStart = (e: React.DragEvent<HTMLDivElement>) => {
+      const dt = e.dataTransfer;
+
+      // Enables dragging of the underlying JSON of nodes into VSCode for
+      // debugging / troubleshooting purposes
+      const resource = JSON.stringify([{
+        resource: '' + c.id,
+        content: JSON.stringify(dragPayload, null, 2),
+        viewState: null,
+        encoding: 'UTF-8',
+        mode: null,
+        isExternal: false,
+      }]);
+
+      dt.setData('CodeEditors', resource);
+      dt.setData('application/x-oli-resource-content', JSON.stringify(dragPayload));
+      dt.setData('text/html', toSimpleText(c as any));
+      dt.setData('text/plain', toSimpleText(c as any));
+      dt.effectAllowed = 'move';
+
+    };
+
+    const handleKeyDown = (e: React.KeyboardEvent) => {
+      if (e.shiftKey && e.key === 'ArrowDown') {
+        onMove(index, false);
+      } else if (e.shiftKey && e.key === 'ArrowUp') {
+        onMove(index, true);
+      }
+    };
+
     return (
-      <ResourceContentFrame
-        key={c.id}
-        allowRemoval={content.size > 1}
-        editMode={editMode}
-        label={label}
-        purpose={c.purpose}
-        purposes={purposes}
-        editingLink={editingLink}
-        onEditPurpose={(purpose: string) => {
-          const u = Object.assign(c, { purpose });
-          props.onEdit(u, index);
-        }}
-        onRemove={onRemove}>
+      <React.Fragment>
+        <DropTarget id={c.id} index={index} onDrop={onDrop}/>
 
-        {editor}
+        <div className="resource-editor"
+          key={c.id}
+          onKeyDown={handleKeyDown}
+          onFocus={e => onFocus(index)}
+          role="option"
+          aria-describedby="content-list-operation"
+          tabIndex={index + 1}>
 
-      </ResourceContentFrame>
+          <div className="resource-content-frame card">
+            <div className="card-header pl-2"
+              draggable={true}
+              onDragStart={handleDragStart}>
+              <div className="d-flex flex-row align-items-baseline">
+                <div className="flex-grow-1">
+                  <DragHandle style={{ height: 24, marginRight: 10 }} /> {label}
+                </div>
+                <Purpose purpose={c.purpose} purposes={purposes}
+                  editMode={editMode} onEdit={(p: string) => onEditPurpose(p)}/>
+                {link}
+                <DeleteButton editMode={content.size > 1} onClick={onRemove}/>
+              </div>
+            </div>
+            <div className="card-body">
+              {editor}
+            </div>
+          </div>
+        </div>
+      </React.Fragment>
     );
   });
 
   return (
-    <div className="d-flex flex-column flex-grow-1">
+    <div className="editors d-flex flex-column flex-grow-1">
       {editors}
+
+      <DropTarget id="last" index={editors.size} onDrop={onDrop}/>
     </div>
   );
 };
