--- assets/src/components/resource/Outline.scss
@@ -1,14 +1,0 @@
-.outline {
-
-  .assistive-text {
-    position: absolute;
-    margin: -1px;
-    border: 0;
-    padding: 0;
-    width: 1px;
-    height: 1px;
-    overflow: hidden;
-    clip: rect(0 0 0 0);
-  }
-
-}
